books_project/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.0.3
kafka-python==2.0.2
djangorestframework-simplejwt==5.2.2
setuptools==75.1.0
PyMySQL==1.1.1
six==1.16.0
